+++ doc/RELEASE-NOTES.md
@@ -8,6 +8,7 @@
 * Added `X509DSADetector` validator for X.509 certificates.
 * Added `ElementBase64WrappingStage`.
 * Added `EmailAddressStringValidator`.
+* Added `FlowConstraint`, `EnableFlow`, `DisableFlow` and `EntityDescriptorFlowConstraintPopulationStage`.
 
 ## Version 0.9.7 ##